Extract part of the checked_partition table in a new check_config one
This new table stores the "configuration" for a scrubber. A
configuration consists in a set of:

  (datastore, object_type, nb_partitions)

This comes with a migration script; WARNING: this script needs to be
checked before deployment on a productiion-sized big database. Any
activity on the database should be stopped before execution.

This is the first step of a series to make the scrubber easier to deploy
on elastic infrastructure.