Add a 'swh scrubber check stalled` command listing stalled partitions
For a given configuration (hence sotrage, object_type and partition scheme)
list partitions that have a start_date but no end_date for a long enough
time.

By default, it will compute the delay for a partition to be considered as
stalled based on the 10 last partitions checked for the given
configuration.