cli: Prevent YAML wrapping in help displayed by Click
The default behavior of Click is to rewrap text based on the width of
the terminal but as a consequence it makes the sample YAML config for
the scrubber displayed in command help quite unreadable as indentation
is lost.

So use \b markers in docstring to ensure proper display of the YAML
config by Click.