cli: Add support for scrubbbing an object storage
Enable to configure and trigger the scrubbing of an object storage with
swh-scrubber CLI, either using partitions of contents provided by a
storage or by consuming the content kafka topic from a SWH journal
(in that case, the --use-journal flag must be provided to "swh check run"
command).

Related to #4694.